Skal skille mellom behandlingstema og behandlingstype ved opprettelse av oppgave etter svar fra kabal.
