+++ src/main/kotlin/no/nav/familie/klage/oppgave/OpprettKabalEventOppgaveTask.kt
@@ -9,9 +9,7 @@ import no.nav.familie.klage.oppgave.OppgaveUtil.lagFristForOppgave
 import no.nav.familie.kontrakter.felles.Behandlingstema
 import no.nav.familie.kontrakter.felles.klage.Fagsystem
 import no.nav.familie.kontrakter.felles.klage.KlageinstansUtfall
-import no.nav.familie.kontrakter.felles.klage.Stønadstype
 import no.nav.familie.kontrakter.felles.objectMapper
-import no.nav.familie.kontrakter.felles.oppgave.Behandlingstype
 import no.nav.familie.kontrakter.felles.oppgave.IdentGruppe
 import no.nav.familie.kontrakter.felles.oppgave.OppgaveIdentV2
 import no.nav.familie.kontrakter.felles.oppgave.OppgavePrioritet
@@ -48,8 +46,6 @@ class OpprettKabalEventOppgaveTask(
 
         val aktivIdent = personRepository.hentAktivIdent(personId)
         val prioritet = utledOppgavePrioritet(opprettOppgavePayload.klageinstansUtfall)
-        val behandlingstype = tilBehandlingstype(opprettOppgavePayload)
-        val behandlingstema = tilBehandlingstema(opprettOppgavePayload, fagsakDomain.stønadstype)
 
         val opprettOppgaveRequest =
             OpprettOppgaveRequest(
@@ -60,28 +56,13 @@ class OpprettKabalEventOppgaveTask(
                 fristFerdigstillelse = lagFristForOppgave(LocalDateTime.now()),
                 beskrivelse = opprettOppgavePayload.oppgaveTekst,
                 enhetsnummer = behandling.behandlendeEnhet,
-                behandlingstema = behandlingstema,
-                behandlingstype = behandlingstype,
+                behandlingstema = opprettOppgavePayload.behandlingstema?.value,
+                behandlingstype = opprettOppgavePayload.behandlingstype,
                 prioritet = prioritet,
             )
 
         val oppgaveId = oppgaveClient.opprettOppgave(opprettOppgaveRequest)
         logger.info("Oppgave opprettet med id $oppgaveId")
-    }
-
-    /**
-     * Dersom behandlingstema er satt i oppgavepayload så er dette behandlingstype som ikke kan brukes i
-     * kombinasjon med behandlingstema - jfr. https://github.com/navikt/oppgave/blob/master/src/main/resources/data/gjelder.json
-     */
-    private fun tilBehandlingstema(opprettOppgavePayload: OpprettOppgavePayload, stønadstype: Stønadstype): String? {
-        return when(opprettOppgavePayload.behandlingstema) {
-            null -> finnBehandlingstema(stønadstype).value
-            else -> null
-        }
-    }
-
-    private fun tilBehandlingstype(opprettOppgavePayload: OpprettOppgavePayload): String? {
-        return opprettOppgavePayload.behandlingstema?.value
     }
 
     companion object {
@@ -93,16 +74,6 @@ class OpprettKabalEventOppgaveTask(
                 TYPE,
                 objectMapper.writeValueAsString(opprettOppgavePayload),
             )
-        }
-    }
-
-    private fun finnBehandlingstema(stønadstype: Stønadstype): Behandlingstema {
-        return when (stønadstype) {
-            Stønadstype.BARNETRYGD -> Behandlingstema.Barnetrygd
-            Stønadstype.OVERGANGSSTØNAD -> Behandlingstema.Overgangsstønad
-            Stønadstype.BARNETILSYN -> Behandlingstema.Barnetilsyn
-            Stønadstype.SKOLEPENGER -> Behandlingstema.Skolepenger
-            Stønadstype.KONTANTSTØTTE -> Behandlingstema.Kontantstøtte
         }
     }
 
@@ -122,4 +93,5 @@ data class OpprettOppgavePayload(
     val fagsystem: Fagsystem,
     val klageinstansUtfall: KlageinstansUtfall?,
     val behandlingstema: Behandlingstema? = null,
+    val behandlingstype: String? = null,
 )
